+++ packages/core/src/link.ts
@@ -1,2 +1,7 @@
+/*
+ * IF YOU CHANGE THIS FILE
+ *    You also need to update the rewrite map in
+ *    packages/babel-preset/src/rewrite-imports.ts
+ */
 export {default as Link} from "next/link"
 export type {LinkProps} from "next/link"
