+++ packages/core/src/router/index.tsx
@@ -6,13 +6,13 @@ import {
   withRouter as withNextRouter,
 } from "next/router"
 import React from "react"
-import {extractRouterParams, useParams} from "./use-params"
+import {extractRouterParams, useParams, useRouterQuery} from "./router-hooks"
 
 export const Router = NextRouter
 export {createRouter, makePublicRouterInstance} from "next/router"
 export {RouterContext} from "next/dist/next-server/lib/router-context"
 
-export {useParam, useParams} from "./use-params"
+export {useParam, useParams, useRouterQuery} from "./router-hooks"
 
 export interface BlitzRouter extends NextRouterType {
   params: ReturnType<typeof extractRouterParams>
@@ -42,46 +42,4 @@ export function useRouter() {
   return React.useMemo(() => {
     return {...router, query, params}
   }, [params, query, router]) as BlitzRouter
-}
-
-export function useRouterQuery() {
-  const router = useNextRouter()
-
-  const query = React.useMemo(() => {
-    const query = decode(router.asPath.split("?")[1])
-    return query
-  }, [router.asPath])
-
-  return query
-}
-
-/*
- * Copied from https://github.com/lukeed/qss
- */
-function toValue(mix: any) {
-  if (!mix) return ""
-  var str = decodeURIComponent(mix)
-  if (str === "false") return false
-  if (str === "true") return true
-  return +str * 0 === 0 ? +str : str
-}
-function decode(str: string) {
-  if (!str) return {}
-  let tmp: any
-  let k
-  const out: Record<string, any> = {}
-  const arr = str.split("&")
-
-  // eslint-disable-next-line no-cond-assign
-  while ((tmp = arr.shift())) {
-    tmp = tmp.split("=")
-    k = tmp.shift()
-    if (out[k] !== void 0) {
-      out[k] = [].concat(out[k], toValue(tmp.shift()) as any)
-    } else {
-      out[k] = toValue(tmp.shift())
-    }
-  }
-
-  return out
 }
